LangChain/requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
# langchain_community
langchain-ollama
#### langserve
langserve
fastapi
uvicorn
sse_starlette
ipykernel
